- fix compilation on mingw (all other platforms treat long as 'long long' only mingw decided to use 'long int' and then fail)

+++ src/sbml/math/FormulaParser.cpp
@@ -398,7 +398,7 @@ LIBSBML_EXTERN
 ASTNode_t *
 SBML_parseFormula (const char *formula)
 {
-  long rule, state, action;
+  long long rule, state, action;
 
   ASTNode_t *node = NULL;
 
@@ -417,7 +417,7 @@ SBML_parseFormula (const char *formula)
 
   while (1)
   {
-    state  = (long) Stack_peek(stack);
+    state  = (long long) Stack_peek(stack);
     action = FormulaParser_getAction(state, token);
 
     if (action == ACCEPT_STATE)
@@ -460,7 +460,7 @@ SBML_parseFormula (const char *formula)
     {
       rule  = -action;
       node  = FormulaParser_reduceStackByRule(stack, rule);
-      state = (long) Stack_peek(stack);
+      state = (long long) Stack_peek(stack);
 
       Stack_push( stack, node );
       Stack_push( stack, (void *) FormulaParser_getGoto(state, rule) );
